requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jiter==0.8.0
Janome==0.5.0
matplotlib==3.10.1
numpy==1.26.4
openai==1.55.3
Pillow==11.1.0
pydantic==2.10.2
pydantic_core==2.27.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
wordcloud==1.9.4
youtube-transcript-api==0.6.3
httplib2>=0.20.4
google-api-python-client>=2.79.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
